Handle pfsConfig file
In addition to a pfsDesign file, this commit enables the script to
work with a pfsConfig file. When `visit` parameter is supplied for
the cases of integer and hex string inputs or `W_VISIT` exists in
the FITS header for a filename input, the script recognizes the input
is a `pfsConfig` file.