Fix an issue for non-example files
For some reason, unnecessary lines were inserted in MaskedPfsDesign.py,
forcing to use an example file and directories. These lines are simpley
removed and it works for general files. I still don't understand why the
piece of code was inserted. Maybe editor's fault or my fault on the copy
and paste. Anyway, thank Mao-san for testing.